requirements.txt
aiohttp
PyYAML
Pillow
